refactor: 패키지명을 gamer로 수정
