+++ src/main/java/domain/analyzer/ResultAnalyzer.java
@@ -2,8 +2,8 @@ package domain.analyzer;
 
 import domain.analyzer.dto.PlayerGameResult;
 import domain.analyzer.dto.ResultAnalysisDto;
-import domain.dealer.Dealer;
-import domain.player.Players;
+import domain.gamer.Dealer;
+import domain.gamer.Players;
 
 import java.util.List;
