+++ src/main/java/domain/gamer/Gamer.java
@@ -1,4 +1,4 @@
-package domain.player;
+package domain.gamer;
 
 import domain.card.CardBundle;
