+++ src/main/java/domain/gamer/dto/PlayerResultDto.java
@@ -1,7 +1,7 @@
-package domain.player.dto;
+package domain.gamer.dto;
 
-import domain.dealer.Dealer;
-import domain.player.Player;
+import domain.gamer.Dealer;
+import domain.gamer.Player;
 
 public record PlayerResultDto(
         PlayerHandDto playerHand,
